+++ src/error.rs
@@ -46,6 +46,8 @@ impl fmt::Display for TwitterErrorCode {
 ///A set of errors that can occur when interacting with Twitter.
 #[derive(Debug)]
 pub enum Error {
+    ///A URL was passed to a shortcut function that didn't match the method being called.
+    BadUrl,
     ///The response from Twitter was formatted incorrectly or in an unexpected manner. The enclosed
     ///values are an explanatory string and, if applicable, the input that caused the error.
     ///
@@ -80,6 +82,7 @@ pub enum Error {
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match *self {
+            Error::BadUrl => write!(f, "URL given did not match API method"),
             Error::InvalidResponse(err, ref ext) => write!(f, "Invalid response received: {} ({:?})", err, ext),
             Error::MissingValue(val) => write!(f, "Value missing from response: {}", val),
             Error::TwitterError(ref err) => write!(f, "Error(s) returned from Twitter: {}", err),
@@ -96,6 +99,7 @@ impl std::fmt::Display for Error {
 impl std::error::Error for Error {
     fn description(&self) -> &str {
         match *self {
+            Error::BadUrl => "URL given did not match API method",
             Error::InvalidResponse(_, _) => "Invalid response received",
             Error::MissingValue(_) => "Value missing from response",
             Error::TwitterError(_) => "Error returned from Twitter",